Named types
Named types assigned with "typename = :type;" and otherwise parsed as a statement, and have scope.
In scope they can appear on the RHS of a typing: "x:typename = ..."
Named types are nomative; basically a subtype of whatever type they name, and can be used to restrict types in variables and function parameters.  Primitives can be named: "gal = :flt" to name a type "gal" which is a subtype of "flt".

Infinite extension of structs changed from "any" to "all" strictly to make dual() monotonic; I still think it should be "any", as extra fields will be ignored so can be of any type.

Declaring a typename also declares a 1-arg constructor function of the same name (and argument type) which wraps a Convert PrimNode.

Clean up some replicated error logic.
Fix bug in TypeTuple with dual not returning self for all self-symmetric parts.
Test optional trailing comment.
Remove done TODO progress notes.