Calls take tuples & unpack them
In support of varargs, calls normally parse as taking exactly 1 argument which is typically a tuple.  The Call eagerly unpacks the tuple (so typically no allocation is made).
Application is now lisp-like with adjacency - "fun arg" is a function application.  Typically arg is surrounded by parens which makes a tuple.  This will mimic C-like calling conventions.  Eg. "sqrt(9)" and "sqrt 9" both work as expected.