heading towards parsing a generic map() call
GCP - update untouched uses of a touched node, even if the touched node did not change types (its all_type() was equal to its final type, common for constants).

Mixed type errors ("a is not a b") fully meet/join, and so can be incrementally updated in both gvn.iter() and gvn.gcp() and still be monotonic.

parse bugs parsing tuples with/without final comma or no initial value (a nulple).

a bunch of missing widenings or sharpenings added

tuple prints how it parses, with '()' not '[]'

error case for passing a tuple where a struct is  needed.  Build a constructor for a struct