WIP more massive hacking about errors
value() calls value_ne() and wraps the result with all input errors.  This is the default error handling, and is overridden by any Node which can ignore input errors - e.g., PhiNode which ignores errors on dead paths.  Value_ne() calls gvn.type_ne() to get the value buried inside an error wrapper (i.e., it ignores the incoming error) and tries to make a sane result, or an error result if the default inputs are in-error.