Overhaul nullable
TypeNil.NIL is a type, but not in the lattice and it breaks the lattice property by allowing types from unrelated type heirarchies to "cross" into each other.  Instead, type-meets take it, but none return it (unless meeting itself).  Thus '0' can be used all contexts where either a number or a pointer are allowed, and keeps type-safety.  A whole lot of null-ness handling and special TypeInt.ZERO/FALSE handling is removed.  Remove TypeUnion (which was only playing the role of the super-nil).

Swap files TypeFun and TypeFunPtr to be closer to what they are.

TypeFunPtr's are not TypeOops.
Swap args to the boilerplate 'free()' call, allowing stronger typing.  Added some more generics.