Simplify cyclic meet
Fix print bug when no type is available.
Fix GVN bug when removing an altered node and not re-inserting until the worklist re-inserts: node is still in graph and should CSE with new instances.  If not, and a new dup is inserted into the GVN table, and the old removed not-re-inserted is deleted, it attempts to remove itself from the GVN table and accidentilly removes the new dup instance.

Add test case for forcing boxing-or-error on a large enough structure.