Back to Plan A...
"becomes" when splitting an alias, but do not change the hash.  Heck-a-lot simpler as the effects after splitting are everywhere instantly ready.  Bits falls back to just being a pile-o-bits which can be and'd/or'd really quickly.  The hash function is weird, to preserve the same-hash both before and after splitting.
Bits always support an up/down notion.  The "_con" field in Bits does not denote a constant (except for nil), but instead is simply a size-compression.
Remove the TypeTree.
No need for a startype since Bits always support an up/down notion.
Normal argument lists in all FunNodes and CallNodes now includes the *memory* argument as a normal argument.  Remove the memory arg as a special arg and move it next to the other args.