Pull memory back out of the args, go back to mostly zero-based args
Rename LibCall to Intrinsic
Rename MergeMem to MemMerge, and add MemSplit as an opposite of a SESE region.
Handle ParmNode -2 for memory index.
TypeMem looses the _any flag.  Most arg-based TypeTuples loose their leading TypeMem.
Intrinsics look a lot of CallNodes, but they are primitives.
Add intrinsics to the TestNode monotonicity test.
Memory is handled more uniformly.
Fix a bunch of cyclic-init errors with BitsAlias splitting bits, but not updating the matching TypeMem.