FunNodes type reset
Inlined primitives get cloned, and their fidx changed which changes the following Unresolved.  Reset all types.
CallNode uses MEET on Unresolved inputs during iter, not JOIN; resolving an input loses the other inputs and must only lift.  Basically, we want the JOIN of inputs... but cannot use the element-wise JOIN because we get all the inputs from a particular Epilog or none of the them.
Passes 2/3rds of testParse2.