testParse2 passes
...and half of testParse3!
Args are not contrapositive, so do a meet.
Generic args are SCALARS not XSCALARS.
TypeFunPtr above is measured from _args._any, and not from the fidxes, which are nearly always a single constant.
Dead epilogs no longer include nil in the type.
When updating a FunNode type, include the cache in the Epilog.