Added TestNode assert that all_type is sane.
node.all_type() has to bound all values returned from node.value() calls.  Still a work-in-progress, failing for NewNode at least.  TestNode tests are rotated to bring the next ones being tested to the top as a debugging-speed help.
Lots more asserts.
Places where args are treated contra-positive, or not, are flagged with QQQ as a TODO.