TypeMemPtr includes mem type
TMP includes the meet of all aliased TypeObjs.  Since TMP includes a TypeObj, no need for BitsAlias to include an error reporting TypeObj.  TypeMem.st no longer takes a seperate TypeObj since the TMP includes one.
NewNode returns a TMP with the Obj included, no need for a multi-return.
Remove buggy optimization when meeting dual with self; it is NOT the case that the meet of dual with self is either dual or self; added TestLattice to verify this.
Remove TFP args do join, again.
Remove a untested.
inline 1-op with no memory effects.
Passes testParse1-2 and most of 3.