Sharpen final/read-only behavior
Cannot cast-away final-ness; cannot cast to a final from a non-final, and read from it.  Cannot cast-away read-only.

NewNode returns both an object and a ptr to it.  Same for all intrinsics with embedded News.  MemMerge merges a TypeObj, not a TMP from a News.  StoreNode has 2 known flavors: storing into all of memory, and storing a specific TypeObj.