Bring back recursive types
Remove TypeName, and but the names on Types directly... since they apply both to TypeObj and most primitives.
Decide I really do need recursive types, and start turning them back on.  Lots of cleanup has happened, so the new version is nicer, but still lots of code to turn on.