peel apart the code ptr from closure ptr
just before stuffing closure ptr in local closure, to be used in linked-list chasing proper closure.
Rename it all CLOSURE instead of DISPLAY (old Pascal term).
Remove cycle-check, as TFPs are never in a cycle with TypeStruct.   Cycles of TS & TMP are fine, and a TS field can contain a TFP, which contains a TS as a closure, but that closure is not itself part of a cycle.  It can contain a TMP to a cycle tho.