WIP attempt to strengthen display
Goal is that display is well-typed inside a function pre-GCP, so can be used to fold up local variables.
Root fail cause is that when the display goes dead, & gets killed - the resulting types in FunPtr and NewObj cannot lift above "sharp-display-dual" or we get monotonicity fails.
Currently using NO_DISP, and need a type representation that is "sharp-disp" and *dead*, and not NIL.
Probably cannot set FunPtr display to null, only to a constant "sharp-display-dead".