Sharpen recursive meets
To LCM up from GCD; fixes approx bug where one side has a cycle of length 1 and the other a cycle of length 2, and the approx is precise but length 2.  The 'isa(approx)' test fails because the meet "crushes" the length-2 cycle to the GCD of length-1.
Fix bug in check_commute not checking since Names added.
Leak a few less Type[] arrays.