Fix lattice bug, cyclic function types
Build & test a lattice which includes Bits and NIL.  Discover NIL essentially expands to [0]->obj and NOT [0]->~obj, as was the prior behavior.  See tests lattice14 and lattice15.
Bits includes a [+/-0].

Include the function return as part of 'args' so that cyclic types which include the return type are discovered, make it the arg[0] so all functions align on where their return is.

NO_DISP is now a TMP of a BitsAlias.NIL to a cyclic struct, so that its "inbetween" GENERIC_FUNCTION and its dual, AND its "same sign as the fidxs".  Dualing preservese the whole sign, so up-prims and down-prims make sense again.

GCP finds cyclic display/function types directly, leading to vastly simpler types (instead of unrolled variations which do/do-not have the function in question in the display).

These changes then touched many things, which all need slight "OBOB" fixes: TFP.nargs now has [0] as return, [1] as the display and [2+] as user args.  CallNode nargs is shifted to match.

WIP making cyclic function/display type directly in the parser.

FunPtrNode loses _tf, gets it from FunNode.
Iter cnt limit assert moves for easier debugging.