Wire when found, flow types when aligned.
CG edges are wired when found - generally when the TFP to a Call finds an actual fidx.  However, the edges might not be aligned in the types (yet), as the unrelated parts of the graph flow types at different speed.  Do not flow values, until the types align - then treat the edges as normal.

Fix reset on DefMem.  Fix code-clone with NewNode to DefMem.  Fix value call on StartMem.