WIP array load
Grammar change: expr: term [binop term]* [postop].
"]" is a post-op, so "ary[idx]" is a call to "[" to make an lval, and a call to "]" with memory and an lval to get the array element.
lvals are a 2-element TypeTuple, and are allowed as an argument everywhere.  Probably this has to change to being a plain 'aa' tuple.

nargs filter both plain funptrs and unresolved.