Add the rest of HM: fresh & occurs_in
Collect the H-M active-scope-stack for the "occurs_in" call in the "fresh" call.
FunPtrs need the current Env to get the active scope.
Excessive comments in CallEpi to help debug.
FunNode makes a simple TVar not a TArgs, and is typed as Ctrl not a structured TVar.  FunNode parms returns self in the [0] ctrl slot.
Replace gvn.con(NO_DISP) with default null in many places.
Assert TArgs does not nest a TArgs in slot 0.
Add (inefficient) fresh & occurs-in code.