Computes 1+2
with the new ValNode and value types.  GCP only, 4 rseeds.
All type names are globally unique; install in a mapping table.
Call & CallEpi special case Value types and ValNode constructors.  Value types can be as arguments, can be bound to a an instance-call and then uses as function arguments (basically curried on the display argument).  "Inlining" the constructor is nothing more than moving the Call args to the Val.
New super-simple call resolution code; based on not allowing 2nd or later args to overlap / ambiguous.
Move FunNode.FUNS to ValFunNode, which allows both ValNodes and FunPtrNodes.  ValNodes behavior "as if" a TMP to a TS for Loads, and "as if" a TFP when used as a Call fcn/display.
More cutout of the (now dead) InstrinsicNode.
Loading a fcn ptr from a ValNode pre-curries the 'self' to the function ptr (basically making an instance-call).
Many minor tweaks and deletes.
TypeObj can now be named; using a named TS indicates a 'value type', and a named 'TypeObj.IS_USED' just means a forward-ref of a 'value type'.
Small tweak to TFP nargs dual calc.
Bugfix in TS.approx1 for displays.