Passes test00
Drop ValFunNode.
Structure named-types: the type-name-space pts to a NewNode, which slot 0 is null, slot 1 is a default instance, and slots 2+ are normal fields.  The default-instance is the short "instance" variant of the type, with all fields at their worst.  Normal proto fields are unchanged: must be final, can be Unresolved or FunPtrs or others.
The value name space refers to the constructor for the type; can be an Unresolved or a FunPtr.
Can do lookups on the type constructor as a proxy for unbound field loads (no TFP loaded; TODO: allow loading unbound TFPs; added syntax for binding).