Remove TypeStruct.approx; does not pass HM
I could not make it both monotonic and not-suck.  See comments in TypeStruct.

In HM, instead never approx.  Keep all pointers simple, and build a sharp pointer at the end.  Since the pointer is always simple, it never updates even as the struct updates; requires lazy discovery of field references to be put on the worklist.

Create a Type parser, for quick bi-directional string<->Type mappings.  Use these in TestHM to replace all the complex deeply nested (and rapidly changing) golden rules with simple String Types.  Requires a recursive deep parsing of types, with dups and Cyclic.install.  Hugely shrinks the golden-rule support.

Cyclic.install handles any Cyclic type, and also an optional Map of Types to be updated with the min types.

Move TypeStruct.sharpen to TypeMem, and use the new Cyclic iterators to make a much simpler & smaller version.