Passes all HM tests
Add a proxy for the Universe of fcns and structs; these can be passed in for any argument.  Escaping results from the program can be passed also.  Non-escaping results never enter the Universe, and will not be called from the Universe.
Apply-lift keeps ptrs shallow.
Sharpen dull ptrs before updating root args, to the nested ones.
Field defaults to ALL BitsAliases for HMT, and this gets refined by GCP, but covers the cases where outside structs are passed in.
Fix bug field names for pairs/triples.