Progress on clz/instance for HM
When HM looks for fields in instances and misses, it looks in a clazz.  If the clazz exists, it must be a final field with a function, whos display is bound to the instance.  When making them during lookup the HM arg-cnt is unknown, so support HM functions with variable args.
Always the clz ends in a ':' in both Env.PROTOs and normal instances.
Replace CANONICAL_INSTANCE with " def".
Replace TV2._is_obj with String TV2._clz.  Instances have _clz set to their clazz.  Clazzes have it empty.