Add a union type
Add a TypeUnion type, to support GCP on Overloads.  Handles 1 each of int,flt,tmp,tfp as a union.  Between TypeNil and the subtypes of TypeNil, so a high TU is below XSCALAR and above e.g. ~int or TypeInt.INT64.dual.  TypeUnion gets an entry in cross-nil and ymeet to support triangulation.  Some fixed Types are pre-made so cross-nil does not need to create a type during a recursive meet.
Slight changes to high TMP & TFP printing.
Slight changes to some hashcodes.