Cleanup bindfp/fp2dsp
BindFP and FP2Dsp are inverts.  Clarify the proto-load sequence: if a Field misses in the current struct, and hits in  the proto, its a "proto load", and if loading a TFP or an overload of TFPs, always Binds the 'this' into the function display.  Change Parser to insert proto loads and binds and resolves for Opers.  Drop the display from FunPtrNode.

Split the display into calls from the function; move the function to the last argument, and have the display as arg#0.  Simplifies a lot of Call value & ideal opts.

Remove add_hm_work, and use the dependency mechanism.
Stronger liveness asserts.
Drop FunNode type splitting (computed now by HM, and will revisit the need for type-specialization later).
Drop Unresolved.