Fix major Fresh bug
Structs track which field inputs are "nongen" and which are "gen".  nongen is zero for actual structs, and only function arguments for closures.  All the rest (all fields in structs and function bodies) are effectively part of a Let/LetRec body and are "generative".
Fresh builds the nongen TV3[] array during Combo init.
Pass a nargs int instead of a is_closure boolean.
Various minor progress bugs.