Passes the bad test case, fails many others.
Cool observation: FP2DSP cannot bypass a Fresh, lest we remove a Fresh on the display - which cans a reference to the call we're about to make.  This passes an UN-fresh'd display to the call with an UN-fresh'd instance of the function, but a FRESH'd function.  These two then unify which undoes the fresh, and prevents polymorphic calls.

Must allow unification of a TVPtr and a TVBase; the Base acts as-if its an empty struct with a single clz pointer and the TVPtr points to a clz.  Allows generic operator unification with unknown LHS.

PrimNode rename make_tvar as wrap_base.  Don;t wrap the bases, but hook more places where we might.

Bring back pinned fields.