Compute Prim types once and freeze them; they do not go on the GVN worklists
Root is reset between tests; replace ROOT with KEEP_ALIVE in many places.
Root takes an extra arg for Prim memory.
Restructure top-level launcher a bit, to reset root.
Add an official Node walker; move a bunch of walks into it.
Field needs Load memory to do a CLZ lookup.  STILL BUSTED HERE.
TMPs of constant aliases are constant, even if the embedded struct is not.
NewNode takes a flag to make a constant.
This makes wrapped primitives constants, and suitable for ConNode.
Change printouts of wrapped primitives to be nicer.
Bring back TVBase, break it out of TVPtr.
Wrapped primitives are a TVStruct, with a CLZ fld to a TVPtr (with alias) to the CLZ struct, and a PRIM fld with the Base.
Fix licsence of NBHML (original copy, not Apache copy).