Progress on the AA/EXE split
Shuffle args on fresh_unify to make it trivial to swap with fresh.unify.  Rename NewNodes more suggestively.
Remove a FP2DSP questionable peephole.
Add a the current Env to the nongen set.  This is broken either way, as we really need a *partial* add; the current Env belongs in the nongen, but completely defined vars belong *out* of it.

Pretty sure I can now remove do_delay_fresh